Initial commit
Just a short bit of code I worked on last night. Copied the bits I needed out of the sat tracker program and pasted them in here along with some parser stuff. The more I adapted the code to work from the command line the worse it got in terms of controlling everything. When its all said and done, each invocation of the program is going to need at minimum 3 arguments but for best accuracy it needs 4 min, and you'll probably not want to use the default elevation and timeframe limits too. So basically you'll be giving 6 arguments every time you use the app, most of which never change.

I've decided the best way to handle this is to add a config option where you can set the command-line arguments there permanently and only have to pass satellite names. Users can still override the config with command-line options, and the program will be operable without any config at all. The config will just make it easier to use. For now the first version will just be command-line arguments.

This code is not actually working btw. Not even close.