Create satpasslist.conf
Sample config file format. Nothing fancy, argument names and values with an equal sign. Anything defined in the config file will be used as the 'base' configuration state with given command-line arguments overriding anything in the config. This allows you to shorten up the command statement significantly. If the value is blank that arg is ignored in the config file.