Config file loading
Added a function that loads and combines the config data from the config file and from the given arguments. Any unset keys in the config file are ignored and anything given on the command-line overrides the config file values. If you have anything weird in your config it will crash, so don't set values to weird things and don't put stuff that doesn't belong into the config.