Fixed paths for config and TLE files
Previously it would try to load the TLE data and the config file from whatever directory the python file was invoked from. When the py file is in your system PATH, it still tries to load the data from whatever folder you happen to be in. Changed that to try to load the config and weather data from the same folder the py file is in. Probably should use local storage but thats more difficult to handle for both windows and linux without a bunch more code. For now under linux there will be a couple files saved into your bin folder: weather.txt and the config file if you need it. Previously it would error out if you didn't have the config file, now it just ignores it if you don't have it.

Also fixed a bug where if the first time you run satpasslist.py you use the --satlist arg, it would previously crash because it couldnt find the TLE data file. Now it makes sure the file exists and is up-to-date before doing anything that needs the list.