Consolidate code
Removed the extra unneeded code for creating a list of satellite names. Also lowered the satname match ratio to 0.90 from 0.95 as even with only a single letter missing or incorrect the matches were between 0.85 and 0.93. Also now using the satellite name returned inside satparams.satellite_name instead of the user supplied name, in case we had to do any corrections. Otherwise the wrong name shows up in the generated list.