Manually set start time for pass predictions
Sometimes you want to know what the schedule was for yesterday, or a week from now. You can now enter a unix epoch timestamp with the option -s (or --starttime) to set the start time for the prediction schedule. Just make sure your TLE data is accurate at that date and time. Also fixed a small bug/omission where I forgot to bound check the input for the elevation limit, which is now fixed to between 0 and 90 degrees. Also added a warning filter to suppress the warning generated by the pyorbital module about tz info for numpy datetime64 objects.