Incorrect east/west code
For some reason, in the code that determines if the pass is on the east or west side of the antenna, I am comparing against the satellite's azimuth at max elevation instead of the longitude like I should be. The weird thing is it produces mostly correct results for my location, and I'm still trying to understand why (crazy maths probably). Using the correct longitude for comparison now so it should always be correct. Tested with both positive and negative GPS longitudes and it agrees with WXtoIMG for all the passes I've checked so far.